refactor[#13] : features 레이어 생성
- setting , snippet 파트를 합칠 widgets 레이어를 생성 할 예정이기 때문
