+++ app/src/features/snippet/ui/CodeSnippet.tsx
