+++ app/src/features/snippet/ui/CopySnippetToClipboardButton.tsx
