+++ app/src/features/snippet/ui/InvisibleSnippetTextArea.tsx
